+++ conda/exports.py
@@ -26,10 +26,13 @@ specs_from_url = specs_from_url
 from .cli.conda_argparse import ArgumentParser  # NOQA
 ArgumentParser = ArgumentParser
 
-from .common.compat import PY3, StringIO,  input, iteritems, lchmod, string_types, text_type  # NOQA
-PY3, StringIO,  input, iteritems, lchmod, string_types, text_type = PY3, StringIO,  input, iteritems, lchmod, string_types, text_type  # NOQA
+from .common.compat import PY3, StringIO,  input, iteritems, string_types, text_type  # NOQA
+PY3, StringIO,  input, iteritems, string_types, text_type = PY3, StringIO,  input, iteritems, string_types, text_type  # NOQA
 from .connection import CondaSession  # NOQA
 CondaSession = CondaSession
+
+from .gateways.disk.link import lchmod  # NOQA
+lchmod = lchmod
 
 from .fetch import TmpDownload  # NOQA
 TmpDownload = TmpDownload
@@ -101,7 +104,7 @@ get_local_urls = lambda: list(get_conda_build_local_url()) or []
 load_condarc = lambda fn: conda.base.context.reset_context([fn])
 from .exceptions import PaddingError  # NOQA
 PaddingError = PaddingError
-from .common.compat import CrossPlatformStLink     # NOQA
+from .gateways.disk.link import CrossPlatformStLink  # NOQA
 CrossPlatformStLink = CrossPlatformStLink
 
 from .models.enums import FileMode  # NOQA